requirements.txt
numpy
matplotlib
MDSplus
configobj
pyqt
